+++ CHANGELOG.md
@@ -1,3 +1,2 @@
-- Supporting for the new minecraft version 1.21.11
-# Issue
-- Configuration menu may not be showing for NeoForge.
+- Added: Option to delay chase when zombies find a target.
+- Fixed: May not show configuration screen on NeoForge.